Clineによるリファクタリング
