+++ src/composables/useDeck.ts
@@ -8,28 +8,53 @@ import {
   loadDeckFromLocalStorage,
   saveDeckName,
   loadDeckName,
+  removeDeckCardsFromLocalStorage,
+  removeDeckNameFromLocalStorage,
 } from "../utils/storage";
-import { encodeDeckCode, decodeDeckCode } from "../utils/deckCode";
 import {
   createNaturalSort,
   createKindSort,
   createTypeSort,
 } from "../utils/sort";
-import { logger } from "../utils/logger";
-import { useToast } from "./useToast";
+import { useDeckCode } from "./useDeckCode";
+import { useDeckReset } from "./useDeckReset";
 
 export function useDeck() {
   const deckCards = ref<DeckCard[]>([]);
   const deckName = ref<string>("新しいデッキ");
-  const deckCode = ref<string>("");
-  const importDeckCode = ref<string>("");
-  const isGeneratingCode = ref<boolean>(false);
-  const showDeckCodeModal = ref<boolean>(false);
-  const showResetConfirmModal = ref<boolean>(false);
-  const error = ref<string | null>(null);
 
-  // トースト通知システムを初期化
-  const { showError, showSuccess } = useToast();
+  // ヘルパー関数
+  const setDeckCards = (cards: DeckCard[]) => {
+    deckCards.value = cards;
+  };
+  const resetDeckCards = () => {
+    deckCards.value = [];
+    removeDeckCardsFromLocalStorage();
+  };
+  const resetDeckName = () => {
+    deckName.value = "新しいデッキ";
+    removeDeckNameFromLocalStorage();
+  };
+
+  // 他のコンポーザブルを統合
+  const {
+    deckCode,
+    importDeckCode,
+    isGeneratingCode,
+    showDeckCodeModal,
+    error: deckCodeError,
+    generateAndShowDeckCode,
+    copyDeckCode,
+    importDeckFromCode,
+    setImportDeckCode,
+  } = useDeckCode(deckCards);
+
+  const {
+    showResetConfirmModal,
+    resetDeck,
+    confirmResetDeck,
+    cancelResetDeck,
+  } = useDeckReset(resetDeckCards, resetDeckName);
 
   // ソート関数インスタンス
   const naturalSort = createNaturalSort();
@@ -135,139 +160,6 @@ export function useDeck() {
   };
 
   /**
-   * デッキをリセット（確認ダイアログを表示）
-   */
-  const resetDeck = (): void => {
-    showResetConfirmModal.value = true;
-  };
-
-  /**
-   * デッキリセットの確認を受け取った場合の処理
-   */
-  const confirmResetDeck = (): void => {
-    deckCards.value = [];
-    deckName.value = "新しいデッキ";
-    localStorage.removeItem(STORAGE_KEYS.DECK_CARDS);
-    localStorage.removeItem(STORAGE_KEYS.DECK_NAME);
-    showResetConfirmModal.value = false;
-    showSuccess("デッキをリセットしました");
-  };
-
-  /**
-   * デッキリセットの確認をキャンセル
-   */
-  const cancelResetDeck = (): void => {
-    showResetConfirmModal.value = false;
-  };
-
-  /**
-   * デッキコードを生成・表示
-   */
-  const generateAndShowDeckCode = (): void => {
-    isGeneratingCode.value = true;
-    error.value = null;
-    try {
-      deckCode.value = encodeDeckCode(deckCards.value);
-      logger.debug("生成されたデッキコード:", deckCode.value);
-      logger.debug("デッキカード数:", deckCards.value.length);
-      logger.debug(
-        "デッキ内容:",
-        deckCards.value.map((item) => `${item.card.id} x${item.count}`)
-      );
-      showDeckCodeModal.value = true;
-    } catch (e) {
-      const errorMessage = "デッキコードの生成に失敗しました";
-      logger.error(errorMessage + ":", e);
-      error.value = errorMessage;
-      showError(errorMessage);
-    } finally {
-      isGeneratingCode.value = false;
-    }
-  };
-
-  /**
-   * デッキコードをクリップボードにコピー
-   */
-  const copyDeckCode = async (): Promise<void> => {
-    error.value = null;
-    try {
-      await navigator.clipboard.writeText(deckCode.value);
-      showSuccess("デッキコードをコピーしました");
-    } catch (e) {
-      const errorMessage = "デッキコードのコピーに失敗しました";
-      logger.error(errorMessage + ":", e);
-      error.value = errorMessage;
-      showError(errorMessage);
-    }
-  };
-
-  /**
-   * デッキコードからインポート
-   */
-  const importDeckFromCode = (availableCards: readonly Card[]): void => {
-    error.value = null;
-
-    // 入力検証：空文字列チェック
-    if (!importDeckCode.value || importDeckCode.value.trim() === "") {
-      const warningMessage = "デッキコードが空です";
-      logger.warn(warningMessage);
-      error.value = warningMessage;
-      showError(warningMessage);
-      return;
-    }
-
-    // 入力検証：基本的な形式チェック（カードIDを/で区切った形式）
-    const trimmedCode = importDeckCode.value.trim();
-
-    // 空文字列や無効な文字が含まれていないかチェック
-    if (
-      trimmedCode.includes("//") ||
-      trimmedCode.startsWith("/") ||
-      trimmedCode.endsWith("/")
-    ) {
-      const warningMessage = "無効なデッキコード形式です";
-      logger.warn(warningMessage);
-      error.value = warningMessage;
-      showError(warningMessage);
-      return;
-    }
-
-    try {
-      logger.debug("デッキコードをデコード中:", trimmedCode);
-      logger.debug("利用可能カード数:", availableCards.length);
-      logger.debug(
-        "利用可能カード(最初の5件):",
-        availableCards.slice(0, 5).map((c) => c.id)
-      );
-
-      const importedCards = decodeDeckCode(trimmedCode, availableCards);
-      logger.debug("デコード結果:", importedCards);
-
-      if (importedCards.length > 0) {
-        deckCards.value = importedCards;
-        importDeckCode.value = "";
-        showDeckCodeModal.value = false;
-        showSuccess(
-          `デッキをインポートしました（${importedCards.length}種類のカード）`
-        );
-      } else {
-        const warningMessage =
-          "有効なカードが見つかりませんでした。カードIDが正しいか確認してください。";
-        logger.warn(warningMessage);
-        logger.debug("入力されたデッキコード:", trimmedCode);
-        error.value = warningMessage;
-        showError(warningMessage);
-      }
-    } catch (e) {
-      const errorMessage = "デッキコードの復元に失敗しました";
-      logger.error(errorMessage + ":", e);
-      logger.debug("入力されたデッキコード:", trimmedCode);
-      error.value = errorMessage;
-      showError(errorMessage);
-    }
-  };
-
-  /**
    * ローカルストレージからデッキを初期化
    */
   const initializeDeck = (availableCards: readonly Card[]): void => {
@@ -296,13 +188,6 @@ export function useDeck() {
     deckName.value = name;
   };
 
-  /**
-   * インポート用デッキコードを設定
-   */
-  const setImportDeckCode = (code: string): void => {
-    importDeckCode.value = code;
-  };
-
   return {
     deckCards,
     deckName,
@@ -311,7 +196,7 @@ export function useDeck() {
     isGeneratingCode,
     showDeckCodeModal,
     showResetConfirmModal,
-    error,
+    error: deckCodeError, // useDeckCodeからのエラーを公開
     sortedDeckCards,
     totalDeckCards,
     addCardToDeck,
@@ -323,7 +208,8 @@ export function useDeck() {
     cancelResetDeck,
     generateAndShowDeckCode,
     copyDeckCode,
-    importDeckFromCode,
+    importDeckFromCode: (availableCards: readonly Card[]) =>
+      importDeckFromCode(availableCards, setDeckCards),
     initializeDeck,
     setDeckName,
     setImportDeckCode,
